use long instead of int to allow tours with many trips (#95)
* use long to avoid overflow exceptions for tours with many trips

* change version in pom to ensure fork is used (and not original repo)

* Revert "change version in pom to ensure fork is used (and not original repo)"

This reverts commit ef5755bb8054a7e9a3f504741ead2741977b048b.

* update changelog

+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/mode_chain/DefaultModeChainGenerator.java
@@ -30,7 +30,7 @@ public class DefaultModeChainGenerator implements ModeChainGenerator {
 	final private int numberOfTrips;
 	final private int numberOfModes;
 
-	final private int maximumAlternatives;
+	final private long maximumAlternatives;
 
 	private int index = 0;
 
@@ -38,10 +38,10 @@ public class DefaultModeChainGenerator implements ModeChainGenerator {
 		this.availableModes = new ArrayList<>(availableModes);
 		this.numberOfModes = availableModes.size();
 		this.numberOfTrips = numberOfTrips;
-		this.maximumAlternatives = ArithmeticUtils.pow(numberOfModes, numberOfTrips);
+		this.maximumAlternatives = ArithmeticUtils.pow((long) numberOfModes, numberOfTrips);
 	}
 
-	public int getNumberOfAlternatives() {
+	public long getNumberOfAlternatives() {
 		return maximumAlternatives;
 	}
